+++ src/main/java/br/com/dbserver/bancoapi/model/Conta.java
@@ -14,8 +14,8 @@ public class Conta {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long agencia;
-    private Long conta;
+    private int agencia;
+    private int conta;
     private LocalDate dataCriacaoConta;
     private double saldo;
     @NotNull
